refactor readers import
build_bulk_manager() func was added to allow providing parameters
to the BulkManager class, because extra args to __init__ can't be used
due to the fact that managers are initialized in whatever place
in Django (and your extra args are not provided there).

Also a dict reader factory was created to simplify registration of new
data formats, e. g. json, xml, yaml, etc.